telegram/app/requirements.txt
fastapi
uvicorn[standard]
aiogram==3.*
requests
python-dotenv
asyncpg
sqlalchemy
databases
httpx
psycopg2-binary>=2.9